+++ MekHQ/src/mekhq/campaign/finances/FinancialReport.java
@@ -179,19 +179,21 @@ public class FinancialReport {
                 .collect(Collectors.toList()));
 
         CampaignOptions campaignOptions = campaign.getCampaignOptions();
+        Accountant accountant = new Accountant(campaign);
+
         if (campaignOptions.payForMaintain()) {
-            r.maintenance = campaign.getWeeklyMaintenanceCosts().multipliedBy(4);
+            r.maintenance = accountant.getWeeklyMaintenanceCosts().multipliedBy(4);
         }
         if (campaignOptions.payForSalaries()) {
-            r.salaries = campaign.getPayRoll();
+            r.salaries = accountant.getPayRoll();
         }
         if (campaignOptions.payForOverhead()) {
-            r.overhead = campaign.getOverheadExpenses();
+            r.overhead = accountant.getOverheadExpenses();
         }
         if (campaignOptions.usePeacetimeCost()) {
-            r.coSpareParts = campaign.getMonthlySpareParts();
-            r.coAmmo = campaign.getMonthlyAmmo();
-            r.coFuel = campaign.getMonthlyFuel();
+            r.coSpareParts = accountant.getMonthlySpareParts();
+            r.coAmmo = accountant.getMonthlyAmmo();
+            r.coFuel = accountant.getMonthlyFuel();
         }
 
         r.contracts = r.contracts.plus(
